requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL driver

# Message Queue & Cache
redis==5.0.1

# Object Storage
boto3==1.34.0

# Settings
pydantic==2.5.0
pydantic-settings==2.1.0

# User directories (cross-platform)
platformdirs==4.1.0

# Medical imaging
nibabel==5.1.0
numpy==1.24.3
scipy==1.11.3
matplotlib==3.7.2

# Deep learning
torch==2.0.1
torchvision==0.15.2

# Utilities
python-dateutil==2.8.2
structlog==23.2.0
psutil==5.9.6

# File handling
aiofiles==23.2.1

# Image processing
Pillow==10.1.0

# Data processing
pandas==2.1.3
